10/10

+++ ft_puthexa.c
@@ -6,13 +6,13 @@
 /*   By: abarriga <abarriga@student.42malaga.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 16:48:01 by abarriga          #+#    #+#             */
-/*   Updated: 2022/10/07 17:15:04 by abarriga         ###   ########.fr       */
+/*   Updated: 2022/10/09 13:37:26 by abarriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"ft_printf.h"
 
-unsigned int	ft_puthexa(const char *str, unsigned long long n)
+unsigned int	ft_puthexa(const char *str, unsigned int n)
 {
 	unsigned int	len;
